Jquery SF assignment
In this assignment, you’ll need to add code to app.js that will cause
Ryu to switch to his looking cool pose when the user presses the “x”
key. When the user releases “x”, Ryu should go back to standing still.

To complete this assignment, you’ll need to use the .keydown() method
from jQuery. The image you’ll need for this is at images/ryu-cool.gif
in the project’s folder. Between looking at the API docs for keydown
and consulting StackOverflow for additional advice, you will be able to
gleen how to use keydown. Some hints: when you use the keydown method,
you need to check to see if the key that the user pressed down is the
specific one you’re looking for. The key code for “x” is 88.

That’s all the guidance we’re going to give you about the
jQuery/JavaScript portion of this assignment—the challenge here is for
you to apply the ideas you learned about showing and hiding divs (and
making divs with background images) in the previous three assignments
to implement the x-key functionality. Although you’re still early in
you’re career as a front end web developer, you’ve got enough
experience at this point that you should feel comfortable beginning to
teach yourself new things, like the keydown() method for jQuery.
Indeed, the ability to rapidly learn new programming techniques is a
core skill for professional programmers.

In addition to the interactive features described above, you should
provide some text in the .main container div that tells users how to
control Ryu. Ideally, this text should be located within the white
.main container.

When you’ve completed this assignment, be sure to add and commit your
changes to Git. You should also publish this page using Github Pages,
and be sure to share a link to your project with your fellow students
and mentor.